Refactor spotlight files into MVC-style folders
