+++ app/scripts/spotlight/core/spotlight-controller.ts
