+++ app/scripts/spotlight/index.ts
@@ -1,1 +1,1 @@
-export { initSpotlight, closeSpotlight } from './ui';
+export { initSpotlight, closeSpotlight } from './ui/spotlight';
