+++ app/scripts/spotlight/types/index.ts
