+++ app/scripts/spotlight/utils/toast.ts
@@ -11,14 +11,3 @@ export function showToast(message: string, doc: Document = document) {
     toast.addEventListener('transitionend', () => toast.remove());
   }, 2000);
 }
-
-export function el<K extends keyof HTMLElementTagNameMap>(
-  tag: K,
-  className?: string,
-  text?: string
-): HTMLElementTagNameMap[K] {
-  const node = document.createElement(tag);
-  if (className) node.className = className;
-  if (text) node.textContent = text;
-  return node;
-}
